Extract WS message handlers to lib/protocol/handlers/ behind a dispatch table
Verbatim move of 26 message types (~850 lines) into six modules by message
family (workspace, conversations, history, team, files, process-control)
plus a dispatch index. The composition root builds one context object whose
member list is frozen by the decomposition spec (processes, clients,
pendingPermissions, transitions, store, agents, workspace, runtime, signals,
broadcast, config); handlers reach root-owned capabilities through it and
lib modules by direct require, never the root.

Four handlers stay in the root deliberately: chat is the kill-window chat
shim (the pinned follow-up write-condition block and every
scheduleScopeReturnKill call site stay in server.js so the kill-window and
legacy-spawn source scans stay valid unretargeted), delegate/end_delegation
are delegation glue and move with the engine, and flush_buffer drains
safeSend's own reconnect buffer.

Conversions, all count-asserted: workspace reads at use time via
getWorkspace(); root capabilities through ctx groups; the search-engine
sites take the engine from ctx.store.ensureSearchEngine() locally.

Four source-scan retargets (Class B, enumerated): pick_folder execFile scan,
get_conversations single-write scan, nested-cancel walk scan, save_agent
cache-order scan now read their lib files. wsHandlerContext/wsDispatch join
_internal by identity for the seam pins.

server.js: 4,957 -> 4,107 lines. Suite 1,473 green.
