+++ server.js
@@ -2224,6 +2224,42 @@ function handleDelegation(msg, processes) {
   if (delegateProc) delegateProc.on('close', handleDelegateClose);
 }
 
+// ===== WS DISPATCH (lib/protocol/handlers/) =====
+// Every WS message type except the four root shims (chat, delegate,
+// end_delegation, flush_buffer) is handled in lib/protocol/handlers/ and
+// selected from this table. The context object's member list is frozen by
+// the decomposition spec: handlers reach root-owned capabilities through it
+// (and lib modules by direct require); they never require the root.
+const protocolHandlers = require('./lib/protocol/handlers/index.js');
+const wsDispatch = protocolHandlers.buildDispatch();
+const wsHandlerContext = {
+  processes: chatProcesses,                       // identity: live process map
+  clients: connectedClients,                      // identity: connected socket set
+  pendingPermissions: pendingPermissionRequests,  // identity: pending permission map
+  // Kill-window machine: stays in the root, injected as a capability
+  // (unused by the slice-9 handlers; the delegation glue consumes it).
+  transitions: { bufferChatIfTransitioning, beginConvoTransition, endConvoTransition, scheduleScopeReturnKill },
+  // Conversation/search query surface: the engine and its caches are root-owned.
+  store: {
+    countConversationMessages, parseSessionHistory,
+    ensureSearchEngine, getSearchEngine: () => searchEngine,
+    clearSearchFailure: () => { searchEngineFailedWorkspace = null; },
+    reconcileSearchBeforeQuery, grepSearchTranscripts, runUniversalSearch,
+  },
+  // Roster/team capabilities: the cache cascade and skill discovery are root-owned.
+  agents: { discoverSkills, invalidateAgentCache, flagRosterRefresh, maybeCompleteSetup, validateAgentSlug, armAgentsDirWatcher },
+  // Workspace lifecycle, boundary guards, and file caches.
+  workspace: {
+    setWorkspaceRoot, healWorkspaceIfMoved, saveRecentWorkspace, loadRecentWorkspaces,
+    discoverWorkspaces, isInsideWorkspace, isSafeCreatePath, getFileTreeCached,
+    invalidateFileListCache, invalidateFileTreeCache, watchOpenFile,
+  },
+  runtime: { getRuntimeStatus, killAllChildren, cleanOrphanedProcesses },
+  signals: { reportStartup, phaseTimer },         // startup telemetry (recordEvent is lib-owned)
+  broadcast: safeSend,
+  config,                                         // lib/config module: getWorkspace at use time
+};
+
 wss.on('connection', (ws) => {
   console.log('Client connected');
   connectedClients.add(ws);
@@ -2607,417 +2643,6 @@ wss.on('connection', (ws) => {
         }
       }
 
-      // Permission response: user approved/denied a tool in the browser UI.
-      // Resolves the pending HTTP long-poll from the PreToolUse hook script.
-      if (msg.type === 'permission_response') {
-        const pending = pendingPermissionRequests.get(msg.requestId);
-        if (pending) {
-          clearTimeout(pending.timer);
-          pendingPermissionRequests.delete(msg.requestId);
-          // "Always allow this folder": the user chose a standing grant along
-          // with the approval. Folder-level, this workspace only.
-          if (msg.allow === true && msg.grantDir) addBoundaryGrant(msg.grantDir);
-          if (pending.res) {
-            // Hook-originated request: answer the held HTTP response.
-            pending.res.writeHead(200, { 'Content-Type': 'application/json' });
-            pending.res.end(JSON.stringify({ allow: msg.allow }));
-          } else if (pending.onDecision) {
-            // Server-originated request (e.g. Codex write markers): callback.
-            try { pending.onDecision(msg.allow === true, 'user'); } catch (e) { console.error('[Permission] onDecision threw:', e); }
-          }
-          console.log(`[Permission] convo=${msg.conversationId} requestId=${msg.requestId} decision=${msg.allow ? 'allow' : 'deny'}`);
-          recordEvent('permission', { conv: msg.conversationId, d: { tool: pending.toolName, decision: msg.allow ? 'allow' : 'deny' } });
-        } else {
-          console.warn(`[Permission] No pending request for requestId=${msg.requestId} (expired or already resolved)`);
-        }
-      }
-
-      // ── CANCEL: User interrupts a running agent ────────────
-      if (msg.type === 'cancel') {
-        const convoId = msg.conversationId;
-        const entry = chatProcesses.get(convoId);
-        if (!entry || entry.exited) {
-          console.log(`[Cancel] convo=${convoId} no active process to cancel`);
-        } else if (entry.idle) {
-          console.log(`[Cancel] convo=${convoId} process is idle, nothing to cancel`);
-        } else {
-          console.log(`[Cancel] convo=${convoId} proc=${entry.processId} agent=${entry.agentId} killing`);
-
-          // Auto-deny any pending permission requests for this conversation
-          for (const [reqId, pending] of pendingPermissionRequests) {
-            if (pending.conversationId === convoId) {
-              clearTimeout(pending.timer);
-              pendingPermissionRequests.delete(reqId);
-              try {
-                if (pending.res) {
-                  pending.res.writeHead(200, { 'Content-Type': 'application/json' });
-                  pending.res.end(JSON.stringify({ allow: false, reason: 'cancelled' }));
-                } else if (pending.onDecision) {
-                  pending.onDecision(false, 'cancelled');
-                }
-              } catch (e) {}
-            }
-          }
-
-          // Mark as cancelled so delegation close handlers skip parent restoration
-          entry.cancelled = true;
-          entry.exited = true;
-
-          // Send cancelled event before kill so client gets it before the done event
-          safeSend(JSON.stringify({
-            type: 'system', subtype: 'cancelled',
-            _conversationId: convoId, _processId: entry.processId, _agent: entry.agentId,
-            _toolCalls: entry.toolCalls || [], _turnStartTime: entry.turnStartTime || null
-          }));
-
-          // Stop the active work. Runtime-aware: Codex entries interrupt
-          // their turn on the SHARED app-server (never kill it on a
-          // conversation cancel); Claude entries kill their child process.
-          if (entry.interrupt) {
-            entry.interrupt();
-          } else {
-            try { killProcessTree(entry.process, 'SIGTERM'); } catch (e) {}
-            // Safety net: SIGKILL after 2 seconds
-            setTimeout(() => {
-              try { killProcessTree(entry.process, 'SIGKILL'); } catch (e) {}
-            }, 2000);
-          }
-
-          // If this is a delegate, also kill every parked ANCESTOR. Walk the
-          // full parent chain rather than only orchestratorEntry, which is null
-          // for non-intercepted nested WS-delegate chains and would otherwise
-          // leak the grandparent orchestrator as a live process.
-          if (entry.delegation) {
-            const killParked = (e) => {
-              if (!e || e.exited) return;
-              e.exited = true;
-              e.cancelled = true;
-              if (e.interrupt) {
-                e.interrupt();
-              } else if (e.process) {
-                try { killProcessTree(e.process, 'SIGTERM'); } catch (err) {}
-                setTimeout(() => { try { killProcessTree(e.process, 'SIGKILL'); } catch (err) {} }, 2000);
-              }
-              console.log(`[Cancel] convo=${convoId} also killed parked ancestor agent=${e.agentId}`);
-            };
-            const seen = new Set([entry]);
-            let d = entry.delegation;
-            let depth = 0;
-            while (d && depth++ < 50) {
-              if (d.orchestratorEntry && !seen.has(d.orchestratorEntry)) {
-                seen.add(d.orchestratorEntry);
-                killParked(d.orchestratorEntry);
-              }
-              const parent = d.originalEntry;
-              if (!parent || seen.has(parent)) break;
-              seen.add(parent);
-              killParked(parent);
-              d = parent.delegation;
-            }
-          }
-
-          // Clean up from the map immediately (close handler will also try but we guard with exited flag)
-          chatProcesses.delete(convoId);
-
-          // Send done so client unblocks
-          safeSend(JSON.stringify({
-            type: 'system', subtype: 'done', code: null,
-            _conversationId: convoId, _processId: entry.processId, _agent: entry.agentId
-          }));
-        }
-      }
-
-      if (msg.type === 'get_workspaces') {
-        // Clear stale workspace pointer if the directory no longer exists
-        if (WORKSPACE && !fs.existsSync(WORKSPACE)) {
-          console.log(`[Workspace] Current workspace no longer exists: ${WORKSPACE}`);
-          setWorkspaceRoot(null);
-        }
-        const wsData = {
-          type: 'workspaces',
-          current: WORKSPACE,
-          recent: loadRecentWorkspaces(),
-          discovered: discoverWorkspaces()
-        };
-        if (WORKSPACE) {
-          try { wsData.analysis = analyzeWorkspace(WORKSPACE, discoverAgents()); } catch (e) { console.warn('  Workspace analysis failed:', e.message); }
-          try { const st = readState(); wsData.workspaceMode = st.workspaceMode || 'knowledge'; wsData.setupComplete = !!st.setupComplete; } catch (e) { /* default */ }
-        }
-        ws.send(JSON.stringify(wsData));
-      }
-
-      // Reported by the client once it has finished rendering a freshly opened
-      // workspace. A summary showing every server phase fast and the client slow
-      // redirects an investigation in one line.
-      if (msg.type === 'client_render_time') {
-        const ms = Number(msg.ms);
-        if (Number.isFinite(ms) && ms >= 0) reportStartup(`client render ${Math.round(ms)}ms`);
-      }
-
-      if (msg.type === 'list_workspaces') {
-        ws.send(JSON.stringify({
-          type: 'workspaces',
-          recent: loadRecentWorkspaces(),
-          discovered: discoverWorkspaces()
-        }));
-      }
-
-      if (msg.type === 'set_workspace') {
-        const dir = msg.path;
-        if (dir && fs.existsSync(dir) && fs.statSync(dir).isDirectory()) {
-          // Kill all running processes when switching workspace
-          const startup = phaseTimer();
-          killAllChildren();
-          setWorkspaceRoot(dir);
-          armAgentsDirWatcher();
-          // Before anything reads state that may have come from another path.
-          healWorkspaceIfMoved(dir);
-          // A workspace switch (including re-selecting the same one) is the
-          // retry trigger for a failed search-engine open, and must not
-          // serve the previous workspace's cached file/skill lists.
-          searchEngineFailedWorkspace = null;
-          invalidateAgentCache();
-          loadRoutineState();
-          saveRecentWorkspace(dir);
-          // Clean up orphaned processes from previous sessions in this workspace
-          cleanOrphanedProcesses();
-          startup.mark('prepare');
-
-          // Detect empty workspace before scaffolding (scaffoldWorkspace adds Doc/skills)
-          let agentList = [];
-          try { agentList = discoverAgents(); } catch (e) { console.warn('  Agent discovery failed:', e.message); }
-          startup.mark('agents');
-          const isEmpty = isEmptyWorkspace(dir, agentList);
-
-          // Empty workspace: scaffold default folders and CLAUDE.md
-          let scaffoldError = null;
-          if (isEmpty) {
-            const result = scaffoldDefaults(dir);
-            if (!result.success) scaffoldError = result.error;
-            invalidateAgentCache();
-          }
-
-          try { scaffoldWorkspace(dir); } catch (e) { console.warn('Scaffold warning:', e.message); }
-          startup.mark('scaffold');
-          console.log(`  Workspace changed to: ${WORKSPACE} (empty=${isEmpty})`);
-
-          // Re-discover agents after scaffolding
-          try { agentList = discoverAgents(); } catch (e) { console.warn('  Agent discovery failed:', e.message); }
-
-          // Auto-detect and store workspace mode
-          const state = readState();
-          if (!state.workspaceMode) {
-            state.workspaceMode = detectWorkspaceMode(dir);
-            writeState(state);
-            console.log(`  Workspace mode auto-detected: ${state.workspaceMode}`);
-          }
-
-          let analysis = null;
-          try { analysis = analyzeWorkspace(dir, agentList); } catch (e) { console.warn('  Workspace analysis failed:', e.message); }
-          startup.mark('analyze');
-          ws.send(JSON.stringify({ type: 'workspace_set', path: WORKSPACE, analysis, isEmpty, workspaceMode: state.workspaceMode, setupComplete: !!state.setupComplete, scaffoldError }));
-          ws.send(JSON.stringify({ type: 'agents', agents: agentList }));
-          try { ws.send(JSON.stringify({ type: 'file_tree', tree: getFileTreeCached() })); } catch (e) { console.warn('  File tree failed:', e.message); }
-          startup.mark('tree');
-          reportStartup(`workspace open: ${startup.summary()}`);
-          // Warm the search index off the open path (reconcile-on-open);
-          // ensureSearchEngine also self-heals lazily on first search.
-          setImmediate(() => { try { ensureSearchEngine(); } catch (e) { console.warn('[Search] warm-up failed:', e.message); } });
-        } else {
-          ws.send(JSON.stringify({ type: 'workspace_error', message: 'Directory not found' }));
-        }
-      }
-
-      if (msg.type === 'pick_folder') {
-        // Async execFile (not the blocking sync variant) so the native folder
-        // dialog does not stall the event loop for up to 60s, freezing all
-        // streams, heartbeats and permission long-polls. The args array
-        // also avoids shell parsing.
-        const { execFile } = require('child_process');
-        // KNOWN LIMITATION: concurrent pick_folder requests spawn overlapping osascript dialogs (not serialized). Cosmetic.
-        execFile('osascript',
-          ['-e', 'POSIX path of (choose folder with prompt "Choose a workspace folder")'],
-          { encoding: 'utf-8', timeout: 60000 },
-          (err, stdout) => {
-            if (err) {
-              // User cancelled or osascript failed
-              ws.send(JSON.stringify({ type: 'folder_picked', path: null }));
-              return;
-            }
-            const result = (stdout || '').trim();
-            if (result) {
-              // Remove trailing slash if present
-              const dir = result.endsWith('/') ? result.slice(0, -1) : result;
-              ws.send(JSON.stringify({ type: 'folder_picked', path: dir }));
-            } else {
-              ws.send(JSON.stringify({ type: 'folder_picked', path: null }));
-            }
-          });
-      }
-
-      if (msg.type === 'create_workspace') {
-        const rawName = (msg.name || '').replace(/[\/\\:*?"<>|]/g, '').trim();
-        if (!rawName) {
-          ws.send(JSON.stringify({ type: 'workspace_error', message: 'Please enter a workspace name' }));
-        } else {
-          const home = process.env.HOME || process.env.USERPROFILE || '';
-          const dir = path.join(home, 'Documents', 'Rundock', rawName);
-          try {
-            fs.mkdirSync(dir, { recursive: true });
-            fs.mkdirSync(path.join(dir, '.claude'), { recursive: true });
-            // Kill all running processes when creating/switching workspace
-            killAllChildren();
-            setWorkspaceRoot(dir);
-            armAgentsDirWatcher();
-            loadRoutineState();
-            saveRecentWorkspace(dir);
-
-            // New workspace is always empty: scaffold defaults
-            let scaffoldError = null;
-            const result = scaffoldDefaults(dir);
-            if (!result.success) scaffoldError = result.error;
-
-            try { scaffoldWorkspace(dir); } catch (e) { console.warn('Scaffold warning:', e.message); }
-            console.log(`  Workspace created: ${WORKSPACE}`);
-
-            const agentList = discoverAgents();
-            const analysis = analyzeWorkspace(dir, agentList);
-
-            // Auto-detect and store workspace mode
-            const state = readState();
-            state.workspaceMode = detectWorkspaceMode(dir);
-            writeState(state);
-
-            ws.send(JSON.stringify({ type: 'workspace_set', path: WORKSPACE, analysis, isEmpty: true, workspaceMode: state.workspaceMode, setupComplete: false, scaffoldError }));
-            ws.send(JSON.stringify({ type: 'agents', agents: agentList }));
-            ws.send(JSON.stringify({ type: 'file_tree', tree: getFileTreeCached() }));
-          } catch (e) {
-            ws.send(JSON.stringify({ type: 'workspace_error', message: 'Could not create workspace: ' + e.message }));
-          }
-        }
-      }
-
-      if (msg.type === 'get_agents') {
-        if (!WORKSPACE) { ws.send(JSON.stringify({ type: 'needs_workspace' })); return; }
-        let agentList = [];
-        try { agentList = discoverAgents(); } catch (e) { console.warn('  Agent discovery failed:', e.message); }
-        ws.send(JSON.stringify({ type: 'agents', agents: agentList }));
-      }
-      if (msg.type === 'get_runtime_status') {
-        ws.send(JSON.stringify({ type: 'runtime_status', ...getRuntimeStatus() }));
-      }
-      if (msg.type === 'get_files') {
-        if (!WORKSPACE) return;
-        try { ws.send(JSON.stringify({ type: 'file_tree', tree: getFileTreeCached() })); } catch (e) { console.warn('  File tree failed:', e.message); }
-      }
-      if (msg.type === 'get_skills') {
-        let skillList = [];
-        try { skillList = discoverSkills(); } catch (e) { console.warn('  Skill discovery failed:', e.message); }
-        ws.send(JSON.stringify({ type: 'skills', skills: skillList }));
-      }
-
-      // ===== WORKSPACE MODE =====
-
-      if (msg.type === 'set_workspace_mode') {
-        const mode = msg.mode;
-        if (mode !== 'code' && mode !== 'knowledge') {
-          ws.send(JSON.stringify({ type: 'workspace_error', message: 'Invalid workspace mode' }));
-        } else {
-          try {
-            const state = readState();
-            state.workspaceMode = mode;
-            writeState(state);
-            console.log(`  Workspace mode changed to: ${mode}`);
-            ws.send(JSON.stringify({ type: 'workspace_mode_changed', mode }));
-          } catch (e) {
-            ws.send(JSON.stringify({ type: 'workspace_error', message: 'Could not update workspace mode: ' + e.message }));
-          }
-        }
-      }
-
-      // ===== SESSION PERSISTENCE =====
-
-      if (msg.type === 'get_conversations') {
-        if (!WORKSPACE) return;
-        // Clean up empty conversations (no sessionId means no message was ever sent)
-        // Only remove if older than 5 minutes to avoid race with sessionId assignment
-        const convos = readConversations();
-        const fiveMinAgo = Date.now() - 5 * 60 * 1000;
-        const cleaned = convos.filter(c => c.sessionId || new Date(c.lastActiveAt || c.createdAt).getTime() > fiveMinAgo);
-        let convosChanged = cleaned.length < convos.length;
-        // Reconcile activeAgentId on load. A pointer to a delegatee is stale
-        // ONLY when there is no live process: the orchestrator resumes after a
-        // delegate returns or the conversation goes idle. Skip any conversation
-        // with a live process, whose activeAgentId (a live delegate) is
-        // legitimate and must not be clobbered mid-delegation.
-        for (const c of cleaned) {
-          if (c.activeAgentId && c.activeAgentId !== c.agentId && !chatProcesses.has(c.id)) {
-            c.activeAgentId = c.agentId;
-            convosChanged = true;
-          }
-        }
-        // Persist at most once per load, and only when something changed
-        // (previously wrote unconditionally, up to twice per load).
-        if (convosChanged) writeConversations(cleaned);
-        // Strip markdown formatting for plain-text previews (mirrors frontend stripMd)
-        function stripMdServer(t) {
-          return t
-            .replace(/\*\*(.*?)\*\*/g, '$1')       // bold
-            .replace(/\*(.*?)\*/g, '$1')            // italic *
-            .replace(/_(.*?)_/g, '$1')              // italic _
-            .replace(/~~(.*?)~~/g, '$1')            // strikethrough
-            .replace(/`([^`]+)`/g, '$1')            // inline code
-            .replace(/^#+\s*/gm, '')                // headings
-            .replace(/\[([^\]]+)\]\([^)]+\)/g, '$1') // links
-            .replace(/\[\[([^\]|]+)\|([^\]]+)\]\]/g, '$2') // wikilinks with alias
-            .replace(/\[\[([^\]]+)\]\]/g, '$1')     // wikilinks
-            .replace(/==(.*?)==/g, '$1')             // highlights
-            .replace(/^[\s]*[-*+]\s/gm, '');         // list markers
-        }
-        // Enrich each conversation for sidebar/profile display. Two passes:
-        //   1. messageCount: sum of user/assistant chat-bubble turns across
-        //      every Claude Code session JSONL the conversation touches. This
-        //      is the canonical source: Rundock's own transcript only covers
-        //      messages emitted after appendTranscript started running and is
-        //      partial or missing for older conversations.
-        //   2. lastAgentId / lastMessagePreview: still sourced from the
-        //      transcript, which is the only place the orchestrator/specialist
-        //      attribution is recorded for the last visible turn.
-        for (const c of cleaned) {
-          try { c.messageCount = countConversationMessages(c); }
-          catch (e) { c.messageCount = 0; }
-          try {
-            const transcript = loadTranscript(c.id);
-            if (!transcript || !transcript.length) continue;
-            for (let i = transcript.length - 1; i >= 0; i--) {
-              const entry = transcript[i];
-              if (entry.role === 'agent' && entry.text) {
-                c.lastAgentId = entry.agent || null;
-                c.lastMessagePreview = stripMdServer(
-                  entry.text
-                    .replace(/<!-- RUNDOCK:(?:SAVE|CREATE)_AGENT name=[\w-]+ -->[\s\S]*?<!-- \/RUNDOCK:(?:SAVE|CREATE)_AGENT -->/g, '')
-                    .replace(/<!-- RUNDOCK:SAVE_SKILL name=[\w-]+ -->[\s\S]*?<!-- \/RUNDOCK:SAVE_SKILL -->/g, '')
-                    .replace(/<!--[\s\S]*?-->/g, '')
-                    .replace(/\n/g, ' ')
-                    .replace(/^(\s*\[[^\]]+\]\s*)+/, '')
-                ).trim().substring(0, 80);
-                break;
-              }
-            }
-          } catch (e) { /* preview enrichment is best-effort */ }
-        }
-        const lastActiveConversationId = readState().lastActiveConversationId || null;
-        ws.send(JSON.stringify({ type: 'conversations', conversations: cleaned, lastActiveConversationId }));
-      }
-
-      if (msg.type === 'set_last_active_conversation') {
-        if (!WORKSPACE) return;
-        const state = readState();
-        if (msg.id) state.lastActiveConversationId = msg.id;
-        else delete state.lastActiveConversationId;
-        writeState(state);
-      }
-
       // Client requests buffered messages after it has loaded conversations and state.
       // Skip stream_event and assistant messages since responseText snapshot covers them.
       if (msg.type === 'flush_buffer') {
@@ -3032,55 +2657,6 @@ wss.on('connection', (ws) => {
             if (ws.readyState === 1) ws.send(m);
           }
         }
-      }
-
-      if (msg.type === 'save_conversation') {
-        if (!WORKSPACE || !msg.conversation || !msg.conversation.id) return;
-        const convos = readConversations();
-        const idx = convos.findIndex(c => c.id === msg.conversation.id);
-        // Only persist metadata, never message content
-        const entry = {
-          id: msg.conversation.id,
-          agentId: msg.conversation.agentId,
-          activeAgentId: msg.conversation.activeAgentId || null,
-          sessionId: msg.conversation.sessionId || null,
-          sessionIds: msg.conversation.sessionIds || [],
-          title: msg.conversation.title,
-          status: msg.conversation.status || 'active',
-          pinned: msg.conversation.pinned || false,
-          pinnedAt: msg.conversation.pinnedAt || null,
-          listIds: Array.isArray(msg.conversation.listIds) ? msg.conversation.listIds.filter(x => typeof x === 'string') : [],
-          createdAt: msg.conversation.createdAt || new Date().toISOString(),
-          lastActiveAt: new Date().toISOString()
-        };
-        if (idx >= 0) { convos[idx] = entry; } else { convos.unshift(entry); }
-        // Cap at 100 conversations
-        writeConversations(convos.slice(0, 100));
-      }
-
-      // ── CONVERSATION LISTS: named many-to-many sidebar groupings ──
-      if (msg.type === 'get_lists') {
-        if (!WORKSPACE) return;
-        ws.send(JSON.stringify({ type: 'lists', lists: readLists() }));
-      }
-
-      if (msg.type === 'create_list') {
-        if (!WORKSPACE) return;
-        const name = typeof msg.name === 'string' ? msg.name.trim().slice(0, 60) : '';
-        if (!name) return;
-        const lists = readLists();
-        // Same name twice is a no-op rather than a duplicate pill.
-        if (!lists.some(l => l.name.toLowerCase() === name.toLowerCase())) {
-          lists.push({ id: 'list-' + Date.now().toString(36) + Math.random().toString(36).slice(2, 6), name, createdAt: new Date().toISOString() });
-          writeLists(lists);
-        }
-        ws.send(JSON.stringify({ type: 'lists', lists }));
-      }
-
-      if (msg.type === 'delete_list') {
-        if (!WORKSPACE || typeof msg.id !== 'string') return;
-        deleteListEverywhere(msg.id);
-        ws.send(JSON.stringify({ type: 'lists', lists: readLists() }));
       }
 
       // ── DELEGATION: orchestrator hands off to another agent in the same conversation ──
@@ -3133,440 +2709,14 @@ wss.on('connection', (ws) => {
         }
       }
 
-      if (msg.type === 'delete_conversation') {
-        if (!WORKSPACE || !msg.id) return;
-        const convos = readConversations().filter(c => c.id !== msg.id);
-        writeConversations(convos);
-        // Drop the conversation's rows from the search index (spec: a
-        // deleted conversation no longer appears in results).
-        if (ensureSearchEngine()) {
-          try { searchEngine.removeConversation(msg.id); } catch (e) { /* rebuild covers it */ }
-        }
-        ws.send(JSON.stringify({ type: 'conversation_deleted', id: msg.id }));
-      }
-
-      if (msg.type === 'read_file') {
-        const fullPath = path.resolve(WORKSPACE, msg.path);
-        if (isInsideWorkspace(fullPath) && fs.existsSync(fullPath)) {
-          ws.send(JSON.stringify({ type: 'file_content', path: msg.path, content: readNormalisedFile(fullPath) }));
-          // Watch the now-open file so a change made outside Rundock (Obsidian,
-          // an agent, another tool) pushes a live refresh to this client.
-          watchOpenFile(ws, msg.path, fullPath);
-        }
-      }
-
-      if (msg.type === 'add_to_team') {
-        // Assign the next order number to an available agent
-        const agentList = discoverAgents();
-        const target = agentList.find(a => a.id === msg.agentId);
-        if (target && target.fileName) {
-          const maxOrder = Math.max(0, ...agentList.filter(a => a.order !== null).map(a => a.order));
-          const nextOrder = maxOrder + 1;
-          const filePath = path.join(WORKSPACE, '.claude', 'agents', target.fileName);
-          let content = fs.readFileSync(filePath, 'utf-8');
-          // Add or update order field in frontmatter
-          if (content.match(/^order:\s/m)) {
-            content = content.replace(/^order:\s.*/m, `order: ${nextOrder}`);
-          } else {
-            // Add order after the type field, or after description
-            content = content.replace(/^(type:\s.*)/m, `$1\norder: ${nextOrder}`);
-            if (!content.match(/^order:/m)) {
-              content = content.replace(/^(description:[\s\S]*?)(\n\w)/m, `$1\norder: ${nextOrder}$2`);
-            }
-          }
-          fs.writeFileSync(filePath, content, 'utf-8');
-          // Invalidate before rediscovering: the roster cache was populated
-          // at the top of this handler, so without this the agents message
-          // would answer with the recruit still order-less and the client
-          // would not show the join until a later refresh.
-          invalidateAgentCache();
-          ws.send(JSON.stringify({ type: 'agents', agents: discoverAgents() }));
-        }
-      }
-
-      // ===== AGENT & SKILL CRUD (server-side, bypasses Claude Code's .claude/ protection) =====
-      // Agents: .claude/agents/{name}.md
-      // Skills: .claude/skills/{name}/SKILL.md
-      // Both use the same pattern: SAVE (upsert) and DELETE via WebSocket messages,
-      // triggered by RUNDOCK:SAVE_AGENT / RUNDOCK:SAVE_SKILL markers in agent responses.
-
-      // save_agent: upsert (create or update). Also handles legacy 'create_agent' and 'update_agent'.
-      if (msg.type === 'save_agent' || msg.type === 'create_agent' || msg.type === 'update_agent') {
-        const name = msg.name || msg.agentId;
-        if (!validateAgentSlug(name)) {
-          ws.send(JSON.stringify({ type: 'agent_error', message: 'Invalid agent name. Use lowercase letters, numbers, and hyphens only.' }));
-        } else {
-          const agentsDir = path.join(WORKSPACE, '.claude', 'agents');
-          if (!fs.existsSync(agentsDir)) fs.mkdirSync(agentsDir, { recursive: true });
-          const filePath = path.join(agentsDir, name + '.md');
-          if (!isInsideWorkspace(filePath)) {
-            ws.send(JSON.stringify({ type: 'agent_error', message: 'Invalid path.' }));
-          } else {
-            const existed = fs.existsSync(filePath);
-            fs.writeFileSync(filePath, msg.content, 'utf-8');
-            // For new agents: auto-assign type and order so they go straight to team
-            if (!existed) {
-              let saved = fs.readFileSync(filePath, 'utf-8');
-              const hasType = saved.match(/^type:\s/m);
-              const hasOrder = saved.match(/^order:\s/m);
-              if (!hasType || !hasOrder) {
-                const currentAgents = discoverAgents();
-                const maxOrder = Math.max(0, ...currentAgents.filter(a => a.order !== null).map(a => a.order));
-                if (!hasType && !hasOrder) {
-                  // No type or order: add both after description, else as the
-                  // first keys inside the frontmatter block. The previous
-                  // `^(---\s*$)/m` matched the OPENING fence and prepended the
-                  // keys BEFORE it, corrupting the frontmatter so the declared
-                  // name/role parsed as body. Anchor to the opening fence
-                  // line and insert AFTER it instead.
-                  if (saved.match(/^description:\s/m)) {
-                    saved = saved.replace(/^(description:\s.*)/m, `$1\ntype: specialist\norder: ${maxOrder + 1}`);
-                  } else {
-                    // KNOWN LIMITATION: this anchor skips if a BOM or leading whitespace precedes the opening fence. Low bite.
-                    saved = saved.replace(/^(---[ \t]*\r?\n)/, `$1type: specialist\norder: ${maxOrder + 1}\n`);
-                  }
-                } else if (hasType && !hasOrder) {
-                  // Has type but no order: add order after type
-                  saved = saved.replace(/^(type:\s.*)/m, `$1\norder: ${maxOrder + 1}`);
-                }
-                fs.writeFileSync(filePath, saved, 'utf-8');
-              }
-            }
-            console.log(`[Agent] ${existed ? 'Updated' : 'Created'}: ${name}`);
-            // Tag the confirmation with the agent's runtime so the client can
-            // suffix the created pill for non-default runtimes.
-            const savedRuntime = String(parseAgentFrontmatter(msg.content).runtime || '').toLowerCase() === 'codex' ? 'codex' : 'claude';
-            ws.send(JSON.stringify({ type: 'agent_saved', agentId: name, updated: existed, runtime: savedRuntime }));
-            // Invalidate BEFORE discovering so the broadcast reflects the new
-            // file. A warm (<2s) cache otherwise omits the just-saved agent
-            // from this first roster broadcast.
-            invalidateAgentCache();
-            const updatedAgents = discoverAgents();
-            ws.send(JSON.stringify({ type: 'agents', agents: updatedAgents }));
-            ws.send(JSON.stringify({ type: 'skills', skills: discoverSkills(updatedAgents) }));
-            flagRosterRefresh();
-            if (!existed) maybeCompleteSetup(updatedAgents);
-          }
-        }
-      }
-
-      if (msg.type === 'delete_agent') {
-        const agentList = discoverAgents();
-        const target = agentList.find(a => a.id === msg.agentId);
-        if (!target || !target.fileName) {
-          ws.send(JSON.stringify({ type: 'agent_error', message: `Agent "${msg.agentId}" not found.` }));
-        } else if (target.type === 'platform') {
-          ws.send(JSON.stringify({ type: 'agent_error', message: 'Cannot delete platform agents.' }));
-        } else {
-          const filePath = path.join(WORKSPACE, '.claude', 'agents', target.fileName);
-          if (isInsideWorkspace(filePath)) {
-            fs.unlinkSync(filePath);
-            console.log(`[Agent] Deleted: ${msg.agentId}`);
-            ws.send(JSON.stringify({ type: 'agent_deleted', agentId: msg.agentId }));
-            invalidateAgentCache(); // before discovery so the broadcast omits the deleted agent
-            const updatedAgents = discoverAgents();
-            ws.send(JSON.stringify({ type: 'agents', agents: updatedAgents }));
-            ws.send(JSON.stringify({ type: 'skills', skills: discoverSkills(updatedAgents) }));
-            flagRosterRefresh();
-          }
-        }
-      }
-
-      // save_skill: upsert (create or update) a skill's SKILL.md file.
-      if (msg.type === 'save_skill') {
-        const name = msg.name;
-        if (!validateAgentSlug(name)) {
-          ws.send(JSON.stringify({ type: 'skill_error', message: 'Invalid skill name. Use lowercase letters, numbers, and hyphens only.' }));
-        } else {
-          const skillDir = path.join(WORKSPACE, '.claude', 'skills', name);
-          if (!isInsideWorkspace(skillDir)) {
-            ws.send(JSON.stringify({ type: 'skill_error', message: 'Invalid path.' }));
-          } else {
-            if (!fs.existsSync(skillDir)) fs.mkdirSync(skillDir, { recursive: true });
-            const filePath = path.join(skillDir, 'SKILL.md');
-            const existed = fs.existsSync(filePath);
-            fs.writeFileSync(filePath, msg.content, 'utf-8');
-            console.log(`[Skill] ${existed ? 'Updated' : 'Created'}: ${name}`);
-            ws.send(JSON.stringify({ type: 'skill_saved', skillId: name, updated: existed }));
-            invalidateAgentCache(); // before discovery so the skills broadcast is fresh
-            const updatedAgents = discoverAgents();
-            ws.send(JSON.stringify({ type: 'skills', skills: discoverSkills(updatedAgents) }));
-            flagRosterRefresh();
-          }
-        }
-      }
-
-      if (msg.type === 'delete_skill') {
-        const name = msg.name;
-        if (!validateAgentSlug(name)) {
-          ws.send(JSON.stringify({ type: 'skill_error', message: 'Invalid skill name.' }));
-        } else {
-          const skillDir = path.join(WORKSPACE, '.claude', 'skills', name);
-          if (!isInsideWorkspace(skillDir) || !fs.existsSync(skillDir)) {
-            ws.send(JSON.stringify({ type: 'skill_error', message: `Skill "${name}" not found.` }));
-          } else {
-            fs.rmSync(skillDir, { recursive: true });
-            console.log(`[Skill] Deleted: ${name}`);
-            ws.send(JSON.stringify({ type: 'skill_deleted', skillId: name }));
-            invalidateAgentCache(); // before discovery so the skills broadcast is fresh
-            const updatedAgents = discoverAgents();
-            ws.send(JSON.stringify({ type: 'skills', skills: discoverSkills(updatedAgents) }));
-            flagRosterRefresh();
-          }
-        }
-      }
-
-      // ── CONVERSATION SEARCH: search titles and transcript content ──
-      if (msg.type === 'search_conversations') {
-        // Conversation-only search. No in-repo client sends this today
-        // (the palette's search_universal replaced the sidebar search field);
-        // retained deliberately as a stable WS surface for stale cached
-        // clients and a possible sidebar-search reinstatement, and kept
-        // honest by the integration suite. Results carry the conversation
-        // entry plus matchType/snippet, extended with sessionId/seq anchors
-        // on content hits. Grep fallback covers runtimes without node:sqlite.
-        (async () => {
-          const query = (msg.query || '').toLowerCase().trim();
-          if (!WORKSPACE || !query) {
-            ws.send(JSON.stringify({ type: 'search_results', results: [], query: msg.query }));
-            return;
-          }
-          const convos = readConversations();
-          // First pass: title matches (instant)
-          const titleMatches = convos.filter(c => (c.title || '').toLowerCase().includes(query)).map(c => ({ ...c, matchType: 'title' }));
-          // Second pass: content matches (FTS index, or the legacy jsonl grep)
-          let contentResults = [];
-          if (ensureSearchEngine()) {
-            try {
-              reconcileSearchBeforeQuery();
-              const byId = new Map(convos.map(c => [c.id, c]));
-              // prefix keeps mid-word typing states matching, on par with
-              // the old substring grep ("discoun" must find "discount").
-              contentResults = searchEngine.searchMessages(msg.query, { limit: 50, prefix: true })
-                .filter(h => byId.has(h.conversationId))
-                .map(h => ({
-                  ...byId.get(h.conversationId), matchType: 'content', snippet: h.snippet,
-                  sessionId: h.sessionId, seq: h.seq, matchCount: h.matchCount,
-                }));
-            } catch (e) {
-              console.warn('[Search] FTS query failed, using grep fallback:', e.message);
-              contentResults = await grepSearchTranscripts(msg.query, convos);
-            }
-          } else {
-            contentResults = await grepSearchTranscripts(msg.query, convos);
-          }
-          // Merge: title matches first, then content-only matches (no duplicates)
-          const titleIds = new Set(titleMatches.map(c => c.id));
-          const merged = [...titleMatches, ...contentResults.filter(c => !titleIds.has(c.id))];
-          ws.send(JSON.stringify({ type: 'search_results', results: merged.slice(0, 50), query: msg.query }));
-        })().catch(err => {
-          console.warn('[Search] Error:', err.message);
-          ws.send(JSON.stringify({ type: 'search_results', results: [], query: msg.query }));
-        });
-      }
-
-      if (msg.type === 'search_universal') {
-        // Cmd+K universal palette: one query across files,
-        // conversations, agents, and skills, grouped by type.
-        runUniversalSearch(msg).then(({ groups, recent }) => {
-          ws.send(JSON.stringify({ type: 'search_universal_results', query: (msg.query || '').trim(), reqId: msg.reqId, groups, recent }));
-        }).catch(err => {
-          // Defensive backstop: each corpus inside runUniversalSearch catches
-          // its own failures (degrading to partial results), so a rejection
-          // here is unexpected. `error: true` lets the client distinguish a
-          // genuine failure from a query with no hits.
-          console.warn('[Search] universal error:', err && err.message ? err.message : err);
-          ws.send(JSON.stringify({
-            type: 'search_universal_results', query: (msg.query || '').trim(), reqId: msg.reqId,
-            groups: { files: [], conversations: [], agents: [], skills: [] }, recent: false, error: true,
-          }));
-        });
-      }
-
-      if (msg.type === 'get_session_history') {
-        const { sessionId, sessionIds, conversationId, limit, offset } = msg;
-
-        // Multi-session merge: load JSONL content from all sessions, then use the
-        // conversation transcript as the ordering and attribution authority.
-        // The transcript records the correct interleaved order from live use;
-        // JSONL sessions group messages per-process and can reorder across agents.
-        if (sessionIds && sessionIds.length > 0) {
-          Promise.all(sessionIds.map(async (s) => {
-            const result = await parseSessionHistory(s.sessionId, 999, 0).catch(() => ({ messages: [] }));
-            return result.messages;
-          })).then(allSessions => {
-            const transcript = loadTranscript(conversationId);
-
-            // Build a pool of JSONL messages for content lookup
-            const stripToolSummaries = (s) => (s || '').replace(/^(\[.*?\]\s*)+/s, '').trim();
-            const jsonlPool = [];
-            for (const sessionMsgs of allSessions) {
-              for (const m of sessionMsgs) {
-                // Skip whitespace-only content. Without this filter, an entry
-                // whose content is just a space character falsely matches any
-                // cleanPrefix that contains a space (i.e. virtually all of
-                // them), so real transcript text gets replaced by empty
-                // bubbles. Whitespace entries are artifacts of tool-heavy
-                // assistant turns where parseSessionHistory joined empty
-                // `text` blocks into a single whitespace string.
-                if (!m.content || !m.content.trim()) continue;
-                // Skip internal delegation messages
-                if (m.role === 'user' && (
-                  m.content.startsWith('CONVERSATION SO FAR:') ||
-                  m.content.startsWith('[SYSTEM:') ||
-                  m.content.startsWith('[DELEGATION BRIEF]')
-                )) continue;
-                // Skip ghost bubbles: empty resume artifacts from orchestrator
-                if (m.role === 'assistant' && m.content.trim() === 'No response requested.') continue;
-                jsonlPool.push({ ...m, _used: false });
-              }
-            }
-
-            // If we have a transcript, use it as the ordering authority
-            const merged = [];
-            if (transcript && transcript.length > 0) {
-              const seenUserMsgs = new Set();
-              for (const t of transcript) {
-                const role = t.role === 'user' ? 'user' : 'assistant';
-                const tText = t.text || '';
-
-                // Routing entries: orchestrator turn that was an immediate Agent-tool
-                // call with no prose. Pass through with type so the client preserves
-                // the agent change for divider rendering but skips the chat bubble.
-                if (t.type === 'routing') {
-                  merged.push({ role: 'assistant', content: tText, agentId: t.agent || null, type: 'routing', timestamp: t.timestamp || null });
-                  continue;
-                }
-
-                if (role === 'user') {
-                  const key = tText.substring(0, 200);
-                  if (seenUserMsgs.has(key)) continue;
-                  seenUserMsgs.add(key);
-                  // Find matching JSONL entry for full content
-                  const match = jsonlPool.find(m => !m._used && m.role === 'user' &&
-                    m.content && m.content.substring(0, 200) === key);
-                  if (match) {
-                    match._used = true;
-                    merged.push({ role: 'user', content: match.content, agentId: null, timestamp: match.timestamp || t.timestamp || null });
-                  } else if (tText) {
-                    merged.push({ role: 'user', content: tText, agentId: null, timestamp: t.timestamp || null });
-                  }
-                } else {
-                  // Agent message: match by content prefix (transcript stores ~200 chars)
-                  const cleanPrefix = stripToolSummaries(tText).substring(0, 100);
-                  if (!cleanPrefix) continue;
-                  const match = jsonlPool.find(m => !m._used && m.role === 'assistant' &&
-                    m.content && m.content.trim() && (
-                      m.content.substring(0, 100).includes(cleanPrefix.substring(0, 60)) ||
-                      cleanPrefix.includes(m.content.substring(0, 60))
-                    ));
-                  if (match) {
-                    match._used = true;
-                    merged.push({ role: 'assistant', content: match.content, agentId: t.agent || null, timestamp: match.timestamp || t.timestamp || null });
-                  } else {
-                    // No JSONL match: use transcript text (may be truncated but better than dropping)
-                    const cleanText = stripToolSummaries(tText);
-                    if (cleanText) {
-                      merged.push({ role: 'assistant', content: cleanText, agentId: t.agent || null, timestamp: t.timestamp || null });
-                    }
-                  }
-                }
-              }
-            } else {
-              // No transcript: fall back to JSONL pool in order, deduplicated
-              const seenUserMsgs = new Set();
-              for (const m of jsonlPool) {
-                if (m.role === 'user') {
-                  const key = m.content.substring(0, 200);
-                  if (seenUserMsgs.has(key)) continue;
-                  seenUserMsgs.add(key);
-                }
-                merged.push({ role: m.role, content: m.content, agentId: m.role === 'user' ? null : null, timestamp: m.timestamp || null });
-              }
-            }
-
-            const total = merged.length;
-            const lim = limit || 200;
-            const off = offset || 0;
-            const start = Math.max(0, total - lim - off);
-            const end = Math.max(0, total - off);
-            ws.send(JSON.stringify({
-              type: 'session_history',
-              conversationId,
-              messages: merged.slice(start, end),
-              totalCount: total,
-              hasMore: start > 0
-            }));
-          }).catch(err => {
-            console.warn('[Session history] Multi-session merge error:', err.message);
-            ws.send(JSON.stringify({ type: 'session_history', conversationId, messages: [], totalCount: 0, hasMore: false }));
-          });
-        } else {
-          // Fallback: single session (backward compatible)
-          parseSessionHistory(sessionId, limit || 20, offset || 0).then(result => {
-            ws.send(JSON.stringify({
-              type: 'session_history',
-              conversationId,
-              messages: result.messages,
-              totalCount: result.totalCount,
-              hasMore: result.hasMore
-            }));
-          }).catch(err => {
-            console.warn('[Session history] Parse error:', err.message);
-            ws.send(JSON.stringify({ type: 'session_history', conversationId, messages: [], totalCount: 0, hasMore: false }));
-          });
-        }
-      }
-
-      if (msg.type === 'save_file') {
-        const fullPath = path.resolve(WORKSPACE, msg.path);
-        if (isInsideWorkspace(fullPath)) {
-          fs.writeFileSync(fullPath, msg.content, 'utf-8');
-          // Keep the search index and the title-layer file list fresh on the
-          // save hot path; guarded so an index failure can never affect the
-          // save itself.
-          invalidateFileListCache(); invalidateFileTreeCache();
-          if (ensureSearchEngine()) {
-            try { searchEngine.noteFileSaved(WORKSPACE, msg.path); } catch (e) { /* reconcile catches up */ }
-          }
-          ws.send(JSON.stringify({ type: 'file_saved', path: msg.path }));
-        }
-      }
-      // Create a note, board, or folder from the Files sidebar. Files must not
-      // clobber an existing path; folders are idempotent (mkdir -p). A fresh
-      // file tree is pushed so the sidebar updates without a manual reload.
-      if (msg.type === 'create_path') {
-        const rel = String(msg.path || '').replace(/^\/+/, '');
-        const full = path.resolve(WORKSPACE, rel);
-        if (!rel || !isInsideWorkspace(full) || !isSafeCreatePath(rel)) {
-          ws.send(JSON.stringify({ type: 'create_error', path: rel, reason: 'invalid path' }));
-        } else if (msg.kind !== 'folder' && fs.existsSync(full)) {
-          ws.send(JSON.stringify({ type: 'create_error', path: rel, reason: 'already exists' }));
-        } else {
-          try {
-            if (msg.kind === 'folder') {
-              fs.mkdirSync(full, { recursive: true });
-            } else {
-              fs.mkdirSync(path.dirname(full), { recursive: true });
-              fs.writeFileSync(full, msg.content || '', 'utf-8');
-              invalidateFileListCache(); invalidateFileTreeCache();
-              if (ensureSearchEngine()) { try { searchEngine.noteFileSaved(WORKSPACE, rel); } catch (e) {} }
-            }
-            ws.send(JSON.stringify({ type: 'file_tree', tree: getFileTreeCached() }));
-            ws.send(JSON.stringify({ type: 'path_created', path: rel, kind: msg.kind }));
-          } catch (e) {
-            ws.send(JSON.stringify({ type: 'create_error', path: rel, reason: String((e && e.message) || e) }));
-          }
-        }
-      }
-      // Reveal a workspace path in the OS file manager (macOS only; a no-op
-      // elsewhere). Guarded to the workspace and a fixed command.
-      if (msg.type === 'reveal_in_finder') {
-        const full = path.resolve(WORKSPACE, String(msg.path || ''));
-        if (isInsideWorkspace(full) && process.platform === 'darwin') {
-          try { require('child_process').spawn('open', ['-R', full], { stdio: 'ignore' }); } catch (e) {}
-        }
-      }
+      // Every other message type dispatches to lib/protocol/handlers/ (see
+      // the WS DISPATCH block above). The four handlers above this line stay
+      // in the root deliberately: chat is the kill-window chat shim (the
+      // pinned follow-up write-condition block), delegate/end_delegation are
+      // delegation glue and move with the engine, and flush_buffer drains
+      // safeSend's own reconnect buffer.
+      const dispatchHandler = wsDispatch[msg.type];
+      if (dispatchHandler) dispatchHandler(wsHandlerContext, ws, msg);
     } catch (e) {
       console.error('Error handling message:', e);
     }
@@ -4935,6 +4085,8 @@ module.exports._internal = {
   chatProcesses, convoTranscripts, pendingPermissionRequests,
   agentAutoResumeCount, disconnectBuffer, connectedClients,
   convoTransitions,
+  // WS dispatch seam (identity: tests pin the frozen context shape)
+  wsHandlerContext, wsDispatch,
   incrementAutoResume, resetAutoResume,
   // universal search
   ensureSearchEngine, runUniversalSearch, conversationSessionsForSearch,
